+++ src/main/java/com/Da_Technomancer/crossroads/API/effects/alchemy/AquaRegiaAlchemyEffect.java
@@ -18,7 +18,7 @@ import net.minecraftforge.oredict.OreDictionary;
 public class AquaRegiaAlchemyEffect implements IAlchEffect{
 	
 	@Override
-	public void doEffect(World world, BlockPos pos, double amount, double temp, EnumMatterPhase phase){
+	public void doEffect(World world, BlockPos pos, int amount, double temp, EnumMatterPhase phase){
 		if(amount >= .1D){
 			for(EntityLivingBase e : world.getEntitiesWithinAABB(EntityLivingBase.class, phase == EnumMatterPhase.FLAME ? new AxisAlignedBB(pos, pos.add(1, 1, 1)) : new AxisAlignedBB(pos.add(-1, -1, -1), pos.add(2, 2, 2)), EntitySelectors.IS_ALIVE)){
 				e.attackEntityFrom(AcidAlchemyEffect.ACID_DAMAGE, ((float) (amount * 2D)));
